refactor(sponsor): get by id
