+++ app/sponsor/[sponsorId]/components/sponsor-side-panels/sponsor-side-panels.types.ts
