+++ app/sponsor/[sponsorId]/page.tsx
@@ -2,14 +2,14 @@
 
 import { withSponsorGuard } from "utils/guards/sponsor-guard";
 
-import { SponsorSidePanels } from "app/sponsor/components/sponsor-side-panels/sponsor-side-panels";
-import { SponsorSectionBudget } from "app/sponsor/features/sponsor-section-budget/sponsor-section-budget";
-import { SponsorSectionHistory } from "app/sponsor/features/sponsor-section-history/sponsor-section-history";
-import { SponsorSectionProject } from "app/sponsor/features/sponsor-section-project/sponsor-section-project";
-
 import { Icon } from "components/layout/icon/icon";
 import { Translate } from "components/layout/translate/translate";
 import { Typography } from "components/layout/typography/typography";
+
+import { SponsorSidePanels } from "./components/sponsor-side-panels/sponsor-side-panels";
+import { SponsorSectionBudget } from "./features/sponsor-section-budget/sponsor-section-budget";
+import { SponsorSectionHistory } from "./features/sponsor-section-history/sponsor-section-history";
+import { SponsorSectionProject } from "./features/sponsor-section-project/sponsor-section-project";
 
 function SponsorPage() {
   return (
